+++ precompiles/tests/integration/mod.rs
@@ -2,6 +2,8 @@ mod helpers;
 mod uint_fixtures;
 
 mod curve;
+mod ecdsa_secp256k1;
+mod eddsa_ed25519;
 mod hashes;
 mod prime_fields;
 mod u256;